+++ src/fs/modules.ts
@@ -1,8 +1,8 @@
 import {red} from 'kleur/colors';
-import {Timings} from '@feltcoop/util/timings.js';
-import {UnreachableError} from '@feltcoop/util/error.js';
-import type {Result} from '@feltcoop/util';
-import {printError} from '@feltcoop/util/print.js';
+import {Timings} from '@feltjs/util/timings.js';
+import {UnreachableError} from '@feltjs/util/error.js';
+import type {Result} from '@feltjs/util';
+import {printError} from '@feltjs/util/print.js';
 
 import {loadSourcePathDataByInputPath, loadSourceIdsByInputPath} from '../fs/inputPath.js';
 import type {PathStats, PathData} from './pathData.js';