+++ src/publish.task.ts
@@ -1,9 +1,9 @@
 import {createInterface as createReadlineInterface} from 'readline';
 import {green, bgBlack, cyan, red, yellow} from 'kleur/colors';
-import type {Logger} from '@feltcoop/util/log.js';
-import {UnreachableError} from '@feltcoop/util/error.js';
-import type {Flavored, Result} from '@feltcoop/util';
-import {spawn} from '@feltcoop/util/process.js';
+import type {Logger} from '@feltjs/util/log.js';
+import {UnreachableError} from '@feltjs/util/error.js';
+import type {Flavored, Result} from '@feltjs/util';
+import {spawn} from '@feltjs/util/process.js';
 import {z} from 'zod';
 
 import {rainbow} from './utils/colors.js';