+++ src/utils/packageJson.ts
@@ -1,5 +1,5 @@
 import {join} from 'path';
-import type {Json} from '@feltcoop/util/json.js';
+import type {Json} from '@feltjs/util/json.js';
 
 import type {Filesystem} from '../fs/filesystem.js';
 import {paths, groPaths, isThisProjectGro} from '../paths.js';